requirements.txt
requests
pymongo
python-dotenv
flask
selenium
certifi
pymongo[srv]